Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,16 +8,16 @@
  * Return 0 on success and otherwise if error.
  */
 
-int main(void)
+long unsigned int main(void)
 {
-int num1;
-int num2;
-int num3;
-int item;
+long unsigned int num1;
+long unsigned int num2;
+long unsigned int num3;
+long unsigned int item;
 num1 = 1;
 num2 = 2;
 item = 50;
-printf("%u, %u", num1, num2);
+printf("%lu, %lu", num1, num2);
 while (item > 0)
 {
 num3 = num1 + num2;
